+++ pulse2percept/models/retina/__init__.py
@@ -12,6 +12,7 @@ suprachoroidal devices, plus the retinal spatial base class they share.
     nanduri2012
     beyeler2019
     granley2021
+    ho2018
 
 .. seealso::
 
@@ -23,6 +24,7 @@ from .beyeler2019 import (AxonMapModel, AxonMapSpatial, ScoreboardModel,
                           ScoreboardSpatial)
 from .granley2021 import (BiphasicAxonMapModel, BiphasicAxonMapSpatial,
                           BiphasicScoreboardModel, BiphasicScoreboardSpatial)
+from .ho2018 import Ho2018Model, Ho2018Spatial, Ho2018Temporal
 from .horsager2009 import Horsager2009Model, Horsager2009Temporal
 from .nanduri2012 import (Nanduri2012Model, Nanduri2012Spatial,
                           Nanduri2012Temporal)
@@ -35,6 +37,9 @@ __all__ = [
     'BiphasicAxonMapSpatial',
     'BiphasicScoreboardModel',
     'BiphasicScoreboardSpatial',
+    'Ho2018Model',
+    'Ho2018Spatial',
+    'Ho2018Temporal',
     'Horsager2009Model',
     'Horsager2009Temporal',
     'Nanduri2012Model',